+++ tests/Unit/DependencyInjection/TsfGatekeeperAiExtensionTest.php
@@ -8,6 +8,7 @@ use Codeception\Test\Unit;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Tsf\GatekeeperAiBundle\Command\ValidateCommand;
 use Tsf\GatekeeperAiBundle\DependencyInjection\TsfGatekeeperAiExtension;
+use Tsf\GatekeeperAiBundle\EventListener\DataObjectListener;
 use Tsf\GatekeeperAiBundle\Installer;
 use Tsf\GatekeeperAiBundle\Service\Config\Settings;
 use Tsf\GatekeeperAiBundle\Service\ProposalStore;
@@ -27,6 +28,7 @@ final class TsfGatekeeperAiExtensionTest extends Unit
         self::assertTrue($container->hasDefinition(ValidateCommand::class));
         self::assertTrue($container->hasDefinition(Installer::class));
         self::assertTrue($container->getDefinition(Installer::class)->isPublic());
+        self::assertSame(['pimcore.dataobject.postDelete'], array_column($container->getDefinition(DataObjectListener::class)->getTag('kernel.event_listener'), 'event'));
 
         $config = $container->getDefinition(Settings::class)->getArgument('$config');
         self::assertSame(['title'], $config['classes']['Product']['enrich']);